Edit code with checkstyle
