+++ src/test/java/duke/tasklist/TaskListTest.java
@@ -1,0 +1,36 @@
+package duke.tasklist;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.time.LocalDate;
+import java.util.ArrayList;
+
+import org.junit.jupiter.api.Test;
+
+import duke.dukeexception.DukeException;
+import duke.task.Task;
+import duke.task.deadlines.Deadlines;
+import duke.task.todos.ToDos;
+
+
+
+public class TaskListTest {
+    @Test
+    public void testMark() throws DukeException {
+        ArrayList<Task> tasks = new ArrayList<>();
+        tasks.add(new Deadlines("Return book", LocalDate.parse("2019-12-12")));
+        TaskList taskList = new TaskList(tasks);
+        taskList.mark(0);
+        assertEquals("  [D][X] Return book (by: Dec 12 2019)", taskList.getTask(0).toString());
+    }
+
+    @Test
+    public void testDelete() throws DukeException {
+        ArrayList<Task> tasks = new ArrayList<>();
+        tasks.add(new ToDos("Read book "));
+        tasks.add(new Deadlines("Return book", LocalDate.parse("2019-12-12")));
+        TaskList taskList = new TaskList(tasks);
+        taskList.delete(0);
+        assertEquals("  [D][ ] Return book (by: Dec 12 2019)", taskList.getTask(0).toString());
+    }
+}
